structure project using Blueprint
