+++ benefitsHub/config.py
@@ -1,0 +1,11 @@
+import os
+
+class Config:
+	SECRET_KEY = os.environ.get('SECRET_KEY')
+	SQLALCHEMY_DATABASE_URI = os.environ.get('SQLALCHEMY_DATABASE_URI')
+	# Handle sending emails
+	MAIL_SERVER = 'smtp.gmail.com'
+	MAIL_PORT = 465
+	MAIL_USE_SSL = True
+	MAIL_USERNAME = os.environ.get('EMAIL_USER')
+	MAIL_PASSWORD = os.environ.get('EMAIL_PASS')
