+++ benefitsHub/posts/forms.py
@@ -1,0 +1,10 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField, TextAreaField
+from wtforms.validators import DataRequired
+
+
+class MakePostForm(FlaskForm):
+    """Class to create a new post"""
+    title = StringField('Title', validators=[DataRequired()])
+    content = TextAreaField('Content', validators=[DataRequired()])
+    submit = SubmitField('Post')
