+++ benefitsHub/templates/account.html
@@ -51,7 +51,7 @@
                     {% endif %}
                   </div>
               </fieldset><br>
-              <div class="form-group"">
+              <div class="form-group">
                   {{ form.submit(class="btn btn-outline-info") }}
               </div>
           </form>
