+++ benefitsHub/templates/create_benefit.html
@@ -67,7 +67,7 @@
                    {% else %}
                        <div class="custom-file">
                            {{ form.benefit_image(class="custom-file-input") }}
-                           <label class="custom-file-label" for="benefit_image">Choose file</label>
+                           <label class="custom-file-label" for="benefit_image"></label>
                        </div>
                    {% endif %}
                </div>
