+++ benefitsHub/templates/create_post.html
@@ -33,7 +33,7 @@
                        {% endif %}
                </div>
            </fieldset>
-           <div class="form-group"">
+           <div class="form-group">
                {{ form.submit(class="btn btn-outline-info") }}
            </div>
        </form>
