+++ benefitsHub/templates/explore_benefits.html
@@ -14,7 +14,7 @@
               <h5>Requirement</h5>
               <p>{{ benefit.requirement | safe }}</p>
               <div article-metadata>
-                <a class="mb-3" href="{{ url_for('user_benefits', username=benefit.benefit_created_by) }}">Post By {{ benefit.benefit_created_by }}</a><br>
+                <a class="mb-3" href="{{ url_for('benefits.user_benefits', username=benefit.benefit_created_by) }}">Post By {{ benefit.benefit_created_by }}</a><br>
                 <small class="text-muted">
                   Posted on {{ benefit.benefit_start_date.strftime('%Y-%m-%d') }}<br>
                   Ends on: {{ benefit.benefit_end_date.strftime('%Y-%m-%d') }}
@@ -30,9 +30,9 @@
     {% for page_num in benefits.iter_pages(left_edge=1, right_edge=1, left_current=1, right_current=2) %}
       {% if page_num %}
         {% if benefits.page == page_num %}
-          <a class="btn btn-info mb-4" href="{{ url_for('explore_benefits', page=page_num) }}">{{ page_num }}</a>
+          <a class="btn btn-info mb-4" href="{{ url_for('benefits.explore_benefits', page=page_num) }}">{{ page_num }}</a>
         {% else %}
-          <a class="btn btn-outline-info mb-4" href="{{ url_for('explore_benefits', page=page_num) }}">{{ page_num }}</a>
+          <a class="btn btn-outline-info mb-4" href="{{ url_for('benefits.explore_benefits', page=page_num) }}">{{ page_num }}</a>
         {% endif %}
       {% else %}
         ...
