+++ benefitsHub/templates/layout.html
@@ -6,7 +6,7 @@
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <!-- Custom stylesheet -->
-    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}"}}">
+    <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}"}}>
     {% if title %}
         <title>BenefitsHub - {{ title }}</title>
     {% else %}
@@ -24,22 +24,22 @@
             </button>
             <div class="collapse navbar-collapse" id="navbarToggle">
               <div class="nav mr-auto">
-                <a class="nav-item nav-link" href="{{ url_for('home') }}">Home</a>
-                <a class="nav-item nav-link" href="{{ url_for('about') }}">About</a>
-                <a class="nav-item nav-link" href="{{ url_for('explore_benefits') }}")">Explore Benefits</a>
+                <a class="nav-item nav-link" href="{{ url_for('main.home') }}">Home</a>
+                <a class="nav-item nav-link" href="{{ url_for('main.about') }}">About</a>
+                <a class="nav-item nav-link" href="{{ url_for('benefits.explore_benefits') }}")>Explore Benefits</a>
               </div>
               <!-- Navbar Right Side -->
               <div class="nav justify-content-end">
                   {% if current_user.is_authenticated %}
-                  <a class="nav-item nav-link" href="{{ url_for('new_benefit') }}">Create New Benefit</a>
-                  <a class="nav-item nav-link" href="{{ url_for('new_post') }}">Make A Post</a>
-                  <a class="nav-item nav-link" href="{{ url_for('view_posts') }}">View Posts</a>
-                  <a class="nav-item nav-link" href="{{ url_for('account') }}">Account</a>
-                    <a class="nav-item nav-link" href="{{ url_for('logout') }}">Logout</a>
+                      <a class="nav-item nav-link" href="{{ url_for('benefits.new_benefit') }}">Create New Benefit</a>
+                      <a class="nav-item nav-link" href="{{ url_for('posts.new_post') }}">Make A Post</a>
+                      <a class="nav-item nav-link" href="{{ url_for('posts.view_posts') }}">View Posts</a>
+                      <a class="nav-item nav-link" href="{{ url_for('users.account') }}">Account</a>
+                      <a class="nav-item nav-link" href="{{ url_for('users.logout') }}">Logout</a>
                   {% else %}
-                      <a class="nav-item nav-link" href="{{ url_for('view_posts') }}">View Posts</a>
-                      <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
-                      <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
+                      <a class="nav-item nav-link" href="{{ url_for('posts.view_posts') }}">View Posts</a>
+                      <a class="nav-item nav-link" href="{{ url_for('users.register') }}">Register</a>
+                      <a class="nav-item nav-link" href="{{ url_for('users.login') }}">Login</a>
                   {% endif %}
               </div>
             </div>
@@ -48,7 +48,7 @@
     </header><br><br>
     <main role="main" class="container text-center">
         <div class="row">
-            <div class="col"">
+            <div class="col">
                 {% with messages = get_flashed_messages(with_categories=true) %}
                     {% if messages %}
                         {% for category, message in messages %}
