+++ benefitsHub/templates/login.html
@@ -40,13 +40,13 @@
                     <div class="form-group">
                         {{ form.submit(class="btn btn-outline-info") }}
                         <small class="text-muted ml-2">
-                        <a href="{{ url_for('reset_request') }}">Forgot Password?</a>
+                        <a href="{{ url_for('users.reset_request') }}">Forgot Password?</a>
                     </small>
                     </div>
                 </form>
                 <div class="border-top pt-3">
                     <small class="text-muted">
-                        Need an account? <a href="{{ url_for('register') }}">Sign up</a>
+                        Need an account? <a href="{{ url_for('users.register') }}">Sign up</a>
                     </small>
                 </div>
             </div>
