+++ benefitsHub/templates/register.html
@@ -59,14 +59,14 @@
                     {% endif %}
                 </div>
             </fieldset>
-            <div class="form-group"">
+            <div class="form-group">
                 {{ form.submit(class="btn btn-outline-info") }}
             </div>
         </form>
     </div>
-    <div class="border-top pt-3"">
+    <div class="border-top pt-3">
         <small class="text-muted">
-            Already have an account? <a href="{{ url_for('login') }}" class="ml-2" }}">Sign In</a>
+            Already have an account? <a href="{{ url_for('users.login') }}" class="ml-2">Sign In</a>
         </small>
     </div>
 {% endblock content %}
