+++ benefitsHub/templates/user_benefits.html
@@ -24,9 +24,9 @@
     {% for page_num in benefits.iter_pages(left_edge=1, right_edge=1, left_current=1, right_current=2) %}
       {% if page_num %}
         {% if benefits.page == page_num %}
-          <a class="btn btn-info mb-4" href="{{ url_for('user_benefits', username=user.username, page=page_num) }}">{{ page_num }}</a>
+          <a class="btn btn-info mb-4" href="{{ url_for('benefits.user_benefits', username=user.username, page=page_num) }}">{{ page_num }}</a>
         {% else %}
-          <a class="btn btn-outline-info mb-4" href="{{ url_for('user_benefits', username=user.username, page=page_num) }}">{{ page_num }}</a>
+          <a class="btn btn-outline-info mb-4" href="{{ url_for('benefits.user_benefits', username=user.username, page=page_num) }}">{{ page_num }}</a>
         {% endif %}
       {% else %}
         ...
