+++ benefitsHub/templates/view_posts.html
@@ -4,7 +4,7 @@
         <article class="media content-section">
           <div class="media-body">
             <div class="article-metadata">
-              <a class="mr-2" href="{{ url_for('user_posts', username=post.author) }}">{{ post.author }}</a>
+              <a class="mr-2" href="{{ url_for('users.user_posts', username=post.author) }}">{{ post.author }}</a>
               <small class="text-muted">{{ post.date_posted.strftime('%Y-%m-%d') }}</small>
             </div>
             <h2>{{ post.title }}</h2>
@@ -16,9 +16,9 @@
       {% if page_num %}
         {% if posts.page == page_num %}
           <span class="btn btn-info mb-4">{{ page_num }}</span>
-          <!-- <a class="btn btn-info mb-4" href="{{ url_for('user_posts', username=username, page=page_num) }}">{{ page_num }}</a> -->
+          <!-- <a class="btn btn-info mb-4" href="{{ url_for('users.user_posts', username=username, page=page_num) }}">{{ page_num }}</a> -->
         {% else %}
-          <a class="btn btn-outline-info mb-4" href="{{ url_for('view_posts', username=username, page=page_num) }}">{{ page_num }}</a>
+          <a class="btn btn-outline-info mb-4" href="{{ url_for('posts.view_posts', username=username, page=page_num) }}">{{ page_num }}</a>
         {% endif %}
       {% else %}
         <span class="btn btn-outline-info mb-4">...</span>
