--- pyproject.toml
@@ -1,17 +1,0 @@
-[tool.poetry]
-name = "alx-portfolioproject"
-version = "0.1.0"
-description = ""
-authors = ["Julius Okwe <85571573+JuliusDgenius@users.noreply.github.com>"]
-readme = "README.md"
-
-[tool.poetry.dependencies]
-python = "^3.10"
-pip = "^24.2"
-flask-wtf = "^1.2.1"
-wtforms = "^3.1.2"
-
-
-[build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
