--- replit.nix
@@ -1,9 +1,0 @@
-{pkgs}: {
-  deps = [
-    pkgs.sqlite-interactive
-    pkgs.saw-tools
-    pkgs.python312Packages.pycodestyle
-    pkgs.tree
-    pkgs.neovim
-  ];
-}
