+++ wsgi.py
@@ -1,4 +1,6 @@
-from benefitsHub import app
+from benefitsHub import create_app
+
+app = create_app()
 
 
 if __name__ == '__main__':
